requirements.txt
absl-py==0.8.1
attrs==19.3.0
cachetools==3.1.1
#certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.3
filelock==3.0.12
funcsigs==1.0.2
future==0.18.2
google-auth==1.8.2
google-auth-oauthlib==0.4.1
grpcio==1.25.0
gym==0.15.4
idna==2.8
importlib-metadata==1.3.0
jsonschema==3.2.0
Markdown==3.1.1
more-itertools==8.0.2
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.2.30
packaging==19.2
pluggy==0.13.1
protobuf==3.11.1
psutil==5.6.7
py==1.8.0
py-spy==0.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyglet==1.3.2
pyparsing==2.4.5
pyrsistent==0.15.6
pytest==5.3.1
#PyYAML==5.2
redis==3.3.11
requests==2.22.0
ray==1.0.1
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.3.3
setproctitle==1.1.10
six==1.13.0
tensorboard==2.1.0
#torch==1.3.1
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
scipy==1.3.3
graphviz
imageio
cffi
Cython
tensorflow==1.14.0
